+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add support for `update-ca-trust` on Linux sytems.
+
 ## v1.4.0
 
   - Add `localhost:purge` to delete all certificates.